add method to extract all data as a list of dictionaries from data container
