+++ src/data_holder.py
@@ -1,4 +1,4 @@
-from validation import Number
+from src.validation import Number
 
 import logging
 logging.basicConfig(level=logging.DEBUG, filename='app.log', filemode='w', format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
@@ -24,6 +24,9 @@ class Data:
     def columns(self):
         return self._columns
 
+    def to_dict(self):
+        return {column: self.__dict__[column] for column in self._columns}
+
 
 class DataContainer:
     _type_of_row = Data
@@ -43,6 +46,13 @@ class DataContainer:
         if list(row.__dict__.keys()) != self._columns:
             raise TypeError
 
+    def all_data_to_dict(self):
+        """Returns all data within the data container in the form of a list holding dictionaries."""
+        to_return = []
+        for data in self.data:
+            to_return.append(data.to_dict())
+        return to_return
+
 
 if __name__ == '__main__':
     class Name(Data):
@@ -51,4 +61,10 @@ if __name__ == '__main__':
             "Column 2": Number(minvalue=1, maxvalue=100),
             "Column 3": Number(minvalue=1, maxvalue=100),
         }
+
+    class NameContainer(DataContainer):
+        _type_of_row = Name
     name_row = Name(**{"Column 1":1, "Column 2":1, "Column 3":1})
+    name_row2 = Name(**{"Column 1":68, "Column 2":66, "Column 3":24})
+    container = NameContainer([name_row, name_row2])
+    print(container.all_data_to_dict())
